Attention Map Explorer/requirements.txt
torch
torchvision
torchmultimodal
transformers
typing_extensions
Pillow
matplotlib
streamlit
numpy
